MINIFICPP-1096 fix PutFile/GenerateFlowFileTests (add binary flag)
Issue: The previously failing test had GenerateFlowFile -> PutFile.
When GenerateFlowFile happened to generate a payload that contained a
'\n' character, then during writing it to a file on windows (w/o binary
flag), it got converted to "\r\n" and when checking the length of a
string read from the very same file, but this time opened with the
binary flag, we could read back the "\r\n" and our expectation didn't
match the result.
Fix: make PutFile use the binary flag